+++ src/main/java/com/example/nutriplan_backend/repository/AlimentoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlimentoRepository extends JpaRepository<Alimento, Long> {
+public interface AlimentoRepository extends JpaRepository<Alimento, Integer> {
 }
